% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare.R
\name{compare.model}
\alias{compare.model}
\title{Compare two model}
\usage{
compare.model(obj1, obj2, fpath)
}
\arguments{
\item{obj1}{an Radiomics.out or Nomogram.out object}

\item{obj2}{an Radiomics.out or Nomogram.out object}

\item{fpath}{an output pptx file path}
}
\value{
output the results in fpath
}
\description{
Compare two model
}
