% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Class_definition.R
\name{cv.run.radiomics}
\alias{cv.run.radiomics}
\title{cross-validation training using 100 LGOCV}
\usage{
cv.run.radiomics(object, dt, num_feature)
}
\arguments{
\item{object}{a Radiomics object}

\item{dt}{a training dataset}

\item{num_feature}{the number of feature remained after mRMR}
}
\value{

}
\description{
cross-validation training using 100 LGOCV
}
