% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dataset.nomogram}
\alias{dataset.nomogram}
\title{nomogram feature data.frame}
\format{
An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 227 rows and 8 columns.
}
\usage{
dataset.nomogram
}
\description{
nomogram feature data.frame
}
\keyword{datasets}
