% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dataset.radiomics}
\alias{dataset.radiomics}
\title{radiomics feature data.frame}
\format{
An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 122 rows and 959 columns.
}
\usage{
dataset.radiomics
}
\description{
radiomics feature data.frame
}
\keyword{datasets}
