% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Class_definition.R
\name{figure.nomogram}
\alias{figure.nomogram}
\title{The function to plot nomogram}
\usage{
figure.nomogram(x, dt, fpath, fun.at)
}
\arguments{
\item{x}{a nomogram object}

\item{dt}{the training dataset}

\item{fpath}{the output pptx file}

\item{fun.at}{function values to label on axis}
}
\description{
The function to plot nomogram
}
