% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Radiomics_Classes.R
\name{figure.radiomics,Radiomics-method}
\alias{figure.radiomics,Radiomics-method}
\title{Plot radiomics figures}
\usage{
\S4method{figure.radiomics}{Radiomics}(object, fpath)
}
\arguments{
\item{object}{Radiomics model}

\item{fpath}{an existing pptx file path}
}
\value{
no return.
}
\description{
Plot radiomics figures
}
