% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Radiomics_Classes.R
\name{figure.radiomics}
\alias{figure.radiomics}
\title{a plot generic function}
\usage{
figure.radiomics(object, ...)
}
\arguments{
\item{object}{a Label_Score class}

\item{...}{other arguments}
}
\value{
output to a designated pptx file
}
\description{
a plot generic function
}
