% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Nomogram_Classes.R
\docType{class}
\name{Nomogram-class}
\alias{Nomogram-class}
\title{Nomogram class}
\value{
a nomogram object
}
\description{
Nomogram class
}
\section{Slots}{

\describe{
\item{\code{uni_p_thresh}}{the threshold determined for selecting the features}

\item{\code{fit}}{the model of nomogram}

\item{\code{formula}}{the formula of the Nomoscore}
}}

