% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Class_definition.R
\docType{class}
\name{Nomogram.out-class}
\alias{Nomogram.out-class}
\title{A nomogram output class}
\value{

}
\description{
A nomogram output class
}
\section{Slots}{

\describe{
\item{\code{Label}}{the Label}

\item{\code{Score}}{the nomogram score}

\item{\code{ulogit}}{the univariate logistic regression output}

\item{\code{mlogit}}{the multivariate logistic regression output}
}}

