% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Nomogram_Classes.R
\name{plot_nomogram}
\alias{plot_nomogram}
\title{The function to plot nomogram}
\usage{
plot_nomogram(x, dt, fpath)
}
\arguments{
\item{x}{a nomogram object}

\item{dt}{the training dataset}

\item{fpath}{the output pptx file}
}
\description{
The function to plot nomogram
}
