% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Nomogram_Classes.R
\name{predict.nomogram}
\alias{predict.nomogram}
\title{Using the nomogram model to predict the nomogram score}
\usage{
predict.nomogram(x, ...)
}
\arguments{
\item{...}{the dataset used to predict}

\item{a}{nomogram object}
}
\value{
a 'Nomogram.out' object
}
\description{
Using the nomogram model to predict the nomogram score
}
