% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Radiomics_Classes.R
\name{predict.radiomics,Radiomics-method}
\alias{predict.radiomics,Radiomics-method}
\title{a function to predict the radscore.}
\usage{
\S4method{predict.radiomics}{Radiomics}(object, dt)
}
\arguments{
\item{object}{a radiomics model}

\item{dt}{a data.frame with the same columns with the training set.}
}
\value{
a Label_Score class
}
\description{
a function to predict the radscore.
}
