% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Radiomics_Classes.R
\name{predict.radiomics}
\alias{predict.radiomics}
\title{a generic function to use radiomics model to predict}
\usage{
predict.radiomics(object, ...)
}
\arguments{
\item{object}{a Radiomics class}

\item{...}{other argument}
}
\value{

}
\description{
a generic function to use radiomics model to predict
}
