% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Class_definition.R
\name{run.nomogram}
\alias{run.nomogram}
\title{Train nomogram model}
\usage{
run.nomogram(x, dt)
}
\arguments{
\item{x}{a nomogram object}

\item{dt}{the training dataset}
}
\value{
an updated nomogram object
}
\description{
Train nomogram model
}
