% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Radiomics_Classes.R
\name{run.radiomics,Radiomics-method}
\alias{run.radiomics,Radiomics-method}
\title{Train model function definition}
\usage{
\S4method{run.radiomics}{Radiomics}(object, dt)
}
\arguments{
\item{object}{a Radiomics model}

\item{dt}{the training dataset}
}
\value{
a updated Radiomics class
}
\description{
Train model function definition
}
