% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Radiomics_Classes.R
\name{run.radiomics}
\alias{run.radiomics}
\title{Train model generic function}
\usage{
run.radiomics(object, ...)
}
\arguments{
\item{object}{a Radiomics model}

\item{...}{other arguments}
}
\value{
a updated Radiomics model
}
\description{
Train model generic function
}
