% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Class_definition.R
\name{validate.nomogram}
\alias{validate.nomogram}
\title{validate the nomogram output}
\usage{
validate.nomogram(object, ...)
}
\arguments{
\item{object}{an Nomogram.out object}

\item{...}{}
}
\value{
an updated Nomogram.out object
}
\description{
validate the nomogram output
}
