% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Radiomics_Classes.R
\name{validate.radiomics,Label_Score-method}
\alias{validate.radiomics,Label_Score-method}
\title{Function to generate the ROC curve and performance metrics}
\usage{
\S4method{validate.radiomics}{Label_Score}(object)
}
\arguments{
\item{object}{a Label_Score object}
}
\value{
an updated Label_Score object
}
\description{
Function to generate the ROC curve and performance metrics
}
