% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Radiomics_Classes.R
\name{validate.radiomics}
\alias{validate.radiomics}
\title{Function to generate the ROC curve and performance metrics}
\usage{
validate.radiomics(object)
}
\arguments{
\item{object}{a Radiomics.out object}
}
\value{
an updated Radiomics.out object
}
\description{
Function to generate the ROC curve and performance metrics
}
