% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Nomogram_Classes.R
\name{figure.nomogram}
\alias{figure.nomogram}
\title{The function to plot nomogram}
\usage{
figure.nomogram(x, dt, fpath)
}
\arguments{
\item{x}{a nomogram object}

\item{dt}{the training dataset}

\item{fpath}{the output pptx file}
}
\description{
The function to plot nomogram
}
