% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Radiomics_Classes.R
\name{figure.radiomics}
\alias{figure.radiomics}
\title{Plot radiomics figures}
\usage{
figure.radiomics(object, fpath)
}
\arguments{
\item{object}{a Radiomics object}

\item{fpath}{a pptx file path}
}
\value{

}
\description{
Plot radiomics figures
}
