% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Class_definition.R
\name{predict.nomogram}
\alias{predict.nomogram}
\title{Using the nomogram model to predict the nomogram score}
\usage{
predict.nomogram(x, dt)
}
\arguments{
\item{x}{a Nomogram object}

\item{dt}{the dataset}
}
\value{
a 'Nomogram.out' object
}
\description{
Using the nomogram model to predict the nomogram score
}
