% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Class_definition.R
\name{predict.radiomics}
\alias{predict.radiomics}
\title{A function to predict the radscore}
\usage{
predict.radiomics(object, dt)
}
\arguments{
\item{object}{a Radiomics object}

\item{dt}{the dataset}
}
\value{
a Radiomics.out object
}
\description{
A function to predict the radscore
}
