% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Radiomics_Classes.R
\docType{class}
\name{Radiomics.out-class}
\alias{Radiomics.out-class}
\title{a class output by run.radiomics}
\value{
a 'Radiomics.out' class
}
\description{
a class output by run.radiomics
}
\section{Slots}{

\describe{
\item{\code{Label}}{the referred label}

\item{\code{Score}}{the radscore calculated by the model}

\item{\code{iROC}}{roc curve}

\item{\code{cmat}}{the metric of the model}
}}

