% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Class_definition.R
\name{run.radiomics}
\alias{run.radiomics}
\title{Train radiomics model}
\usage{
run.radiomics(object, dt, num_feature)
}
\arguments{
\item{object}{a Radiomics object}

\item{dt}{the training dataset}

\item{num_feature}{the number of feature remained after mRMR}
}
\value{
a updated Radiomics object
}
\description{
Train radiomics model
}
