% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Radiomics_Classes.R
\name{validate.radiomics}
\alias{validate.radiomics}
\title{a validate generic function}
\usage{
validate.radiomics(object)
}
\arguments{
\item{object}{a Label_Score object}
}
\value{
an updated Label_Score object
}
\description{
a validate generic function
}
